requirements.txt
aniso8601==7.0.0
astroid==2.3.1
autopep8==1.4.4
coverage==4.5.4
Django==2.2.6
django-filter==2.2.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pep8==1.7.1
pluggy==0.13.0
promise==2.2.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pylint==2.4.2
pyparsing==2.4.2
python-dotenv==0.10.3
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tox==3.14.0
typed-ast==1.4.0
virtualenv==16.7.5
wrapt==1.11.2
zipp==0.6.0
